refactor autosize hook for datagrid
