+++ packages/sdk-react/src/components/receivables/CreditNotesTable/CreditNotesTable.tsx
@@ -1,11 +1,11 @@
-import { useEffect, useState } from 'react';
-import * as ReactDOM from 'react-dom';
+import { useState } from 'react';
 
 import { components } from '@/api';
 import { ScopedCssBaselineContainerClassName } from '@/components/ContainerCssBaseline';
 import { InvoiceCounterpartName } from '@/components/receivables/InvoiceCounterpartName';
 import { InvoiceStatusChip } from '@/components/receivables/InvoiceStatusChip';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
+import { useAutosizeGridColumns } from '@/core/hooks/useAutosizeGridColumns';
 import { useCurrencies } from '@/core/hooks/useCurrencies';
 import { useReceivables } from '@/core/queries/useReceivables';
 import { ReceivableCursorFields } from '@/enums/ReceivableCursorFields';
@@ -18,7 +18,7 @@ import { DateTimeFormatOptions } from '@/utils/DateTimeFormatOptions';
 import { t } from '@lingui/macro';
 import { useLingui } from '@lingui/react';
 import { Box } from '@mui/material';
-import { DataGrid, GridSortModel, useGridApiRef } from '@mui/x-data-grid';
+import { DataGrid, GridSortModel } from '@mui/x-data-grid';
 import { GridSortDirection } from '@mui/x-data-grid/models/gridSortModel';
 
 import { ReceivableFilters } from '../ReceivableFilters';
@@ -77,28 +77,7 @@ const CreditNotesTableBase = ({ onRowClick }: CreditNotesTableProps) => {
     setPaginationToken(undefined);
   };
 
-  // Adapted from https://mui.com/x/react-data-grid/column-dimensions/#autosizing-asynchronously
-  // setTimeout and flushSync are necessary for call order control
-  // Docs say:
-  // The Data Grid can only autosize based on the currently rendered cells.
-  // DOM access is required to accurately calculate dimensions
-  const gridApiRef = useGridApiRef();
-  useEffect(() => {
-    setTimeout(() => {
-      ReactDOM.flushSync(() => {
-        setTimeout(() => {
-          if (gridApiRef.current?.autosizeColumns) {
-            // noinspection JSIgnoredPromiseFromCall
-            gridApiRef.current?.autosizeColumns({
-              columns: ['amount'],
-              includeHeaders: true,
-              includeOutliers: true,
-            });
-          }
-        }, 1);
-      });
-    }, 1);
-  }, [gridApiRef, creditNotes]);
+  const gridApiRef = useAutosizeGridColumns(creditNotes);
 
   const className = 'Monite-CreditNotesTable';
