+++ packages/sdk-react/src/components/receivables/InvoicesTable/InvoicesTable.tsx
@@ -1,10 +1,10 @@
-import { useEffect, useState } from 'react';
-import * as ReactDOM from 'react-dom';
+import { useState } from 'react';
 
 import { components } from '@/api';
 import { ScopedCssBaselineContainerClassName } from '@/components/ContainerCssBaseline';
 import { InvoiceStatusChip } from '@/components/receivables/InvoiceStatusChip';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
+import { useAutosizeGridColumns } from '@/core/hooks/useAutosizeGridColumns';
 import { useCurrencies } from '@/core/hooks/useCurrencies';
 import { useReceivables } from '@/core/queries/useReceivables';
 import { ReceivableCursorFields } from '@/enums/ReceivableCursorFields';
@@ -22,7 +22,6 @@ import {
   GridRenderCellParams,
   GridSortDirection,
   GridSortModel,
-  useGridApiRef,
 } from '@mui/x-data-grid';
 
 import { InvoiceCounterpartName } from '../InvoiceCounterpartName';
@@ -99,28 +98,7 @@ const InvoicesTableBase = ({
       'onRowActionClick' in restProps && restProps.onRowActionClick,
   });
 
-  // Adapted from https://mui.com/x/react-data-grid/column-dimensions/#autosizing-asynchronously
-  // setTimeout and flushSync are necessary for call order control
-  // Docs say:
-  // The Data Grid can only autosize based on the currently rendered cells.
-  // DOM access is required to accurately calculate dimensions
-  const gridApiRef = useGridApiRef();
-  useEffect(() => {
-    setTimeout(() => {
-      ReactDOM.flushSync(() => {
-        setTimeout(() => {
-          if (gridApiRef.current?.autosizeColumns) {
-            // noinspection JSIgnoredPromiseFromCall
-            gridApiRef.current?.autosizeColumns({
-              columns: ['amount'],
-              includeHeaders: true,
-              includeOutliers: true,
-            });
-          }
-        }, 1);
-      });
-    }, 1);
-  }, [gridApiRef, invoices]);
+  const gridApiRef = useAutosizeGridColumns(invoices);
 
   const className = 'Monite-InvoicesTable';
